Exception handling

+++ src/main/java/com/lsmartin/dscommerce/services/ProductService.java
@@ -1,17 +1,20 @@
 package com.lsmartin.dscommerce.services;
 
-import java.util.List;
-import java.util.Optional;
-
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.lsmartin.dscommerce.dto.ProductDTO;
 import com.lsmartin.dscommerce.entities.Product;
 import com.lsmartin.dscommerce.repositories.ProductRepository;
+import com.lsmartin.dscommerce.services.exceptions.DataBaseException;
+import com.lsmartin.dscommerce.services.exceptions.ResourceNotFoundException;
+
+import jakarta.persistence.EntityNotFoundException;
 
 @Service
 public class ProductService {
@@ -21,9 +24,9 @@ public class ProductService {
 	
 	@Transactional(readOnly = true)
 	public ProductDTO findById(Long id) {
-		
-		Optional<Product> result = productRepository.findById(id);
-		Product product = result.get();
+
+		Product product = productRepository.findById(id).orElseThrow(
+				() -> new ResourceNotFoundException("Recurso não encontrado"));
 		ProductDTO dto = new ProductDTO(product);
 		return dto;
 		
@@ -46,17 +49,40 @@ public class ProductService {
 	}
 	
 	@Transactional
-	public ProductDTO update(Long id, ProductDTO dto) {			
-		Product entity = productRepository.getReferenceById(id)	;			
-		copyDtoToEntity(dto, entity);
-		entity = productRepository.save(entity);
-		return new ProductDTO(entity);		
+	public ProductDTO update(Long id, ProductDTO dto) {		
+		
+		try {
+			
+			Product entity = productRepository.getReferenceById(id)	;			
+			copyDtoToEntity(dto, entity);
+			entity = productRepository.save(entity);
+			return new ProductDTO(entity);
+			
+		} catch (EntityNotFoundException e) {
+			throw new ResourceNotFoundException("Recurso não encontrado");
+		}
+				
 	}
 	
-	@Transactional
+	@Transactional(propagation = Propagation.SUPPORTS) //
 	public void delete(Long id) {
 		
-		productRepository.deleteById(id);
+		if (!productRepository.existsById(id)) {
+			
+			throw new ResourceNotFoundException("Recurso não encontrado");
+			
+		}
+		
+		try {
+			
+			productRepository.deleteById(id);
+			
+		} catch (DataIntegrityViolationException e) {
+			throw new DataBaseException("Falha de integridade referêncial")	;
+			
+			}
+		
+		
 		
 	}
